+++ Reflect-App/source/reflect/core/src/js/authentication.js
@@ -5,7 +5,7 @@
  * @link https://github.com/davidgoy/reflect
  * @copyright 2020 Min Tat Goy
  * @license https://www.gnu.org/licenses/gpl.html   GPLv2 or later
- * @version 1.0.0-beta.4
+ * @version 1.0.0-beta.5
  * @since File available since v1.0.0-alpha.1
  */
 
@@ -18,6 +18,10 @@ window.addEventListener('DOMContentLoaded', function() {
 
     const siteKeyInput = document.querySelector('#siteKey');
     siteKeyInput.value = '';
+
+    const authenticateButton = document.querySelector('#authenticateButton');
+    const authenticateButtonSpinner = document.querySelector('#authenticateButtonSpinner');
+    const authenticateButtonText = document.querySelector('#authenticateButtonText');
 
 
     //--------------------------------------------------------------------------
@@ -34,6 +38,10 @@ window.addEventListener('DOMContentLoaded', function() {
 
       form.addEventListener('submit', function(event) {
 
+        authenticateButton.setAttribute('disabled', 'disabled');
+        authenticateButtonSpinner.classList.remove('d-none');
+        authenticateButtonText.innerHTML = 'AUTHENTICATING...';
+
         const formData = new FormData(form);
 
         formData.append('doAsync', 'authenticateSiteKey');
@@ -48,6 +56,10 @@ window.addEventListener('DOMContentLoaded', function() {
         */
 
         apiGetData(formData).then(function(response) {
+
+          authenticateButton.removeAttribute('disabled');
+          authenticateButtonSpinner.classList.add('d-none');
+          authenticateButtonText.innerHTML = 'AUTHENTICATE';
 
           if(response !== 'false') {